requirements.txt
awscli==1.16.135
Babel==2.6.0
boto3==1.9.96
botocore==1.12.125
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Delorean==1.0.0
Django==2.2
djangorestframework==3.9.1
docutils==0.14
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.1
humanize==0.5.1
idna==2.8
jmespath==0.9.3
oauthlib==3.0.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
sqlparse==0.3.0
stripe==2.23.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
